token-minted-webhook-example/requirements.txt
fastapi==0.104.0
PyJWT==2.8.0
uvicorn==0.23.2

